Hiring page restructured according to css guidlines
