+++ src/Containers/Recruitment.js
@@ -2,9 +2,11 @@ import React, { Component } from "react";
 
 //import components here
 import RecruitmentComponent from "../Components/recruitmentComponent";
-// import Footer from "./Footer";
+import Footer from "./Footer";
 import Navbar from "../Containers/Navbar"
+import Cards from "../Components/Cards";
 import "../Stylesheets/Navbar.css";
+
 
 class Recruitment extends Component {
 	render() {
@@ -12,6 +14,8 @@ class Recruitment extends Component {
 			<React.Fragment>
 				<Navbar/>
 				<RecruitmentComponent />
+				
+				<Footer/>
 			</React.Fragment>
 		);
 	}
