+++ src/Containers/team.js
@@ -1,6 +1,7 @@
 import React, { Component } from "react";
 import TeamComponent from "../Components/teamComponent";
 import Navbar from "./Navbar";
+import Footer from "./Footer"
 
 class Team extends Component {
 	render() {
@@ -8,6 +9,7 @@ class Team extends Component {
 			<>
 				<Navbar />
 				<TeamComponent />
+				<Footer/>
 			</>
 		);
 	}
