+++ src/Stylesheets/Recruitment.css
@@ -1,16 +1,18 @@
 @import url("https://fonts.googleapis.com/css2?family=Caveat&display=swap");
 .Hiringcontainer {
 	display: flex;
+	flex-direction: column;
 	background-color: #121221;
-	height: 100vh;
+	/* height: 100vh; */
 	width: 100%;
 	top: 0;
 	bottom: 0;
 	left: 0;
 	right: 0;
 	position: relative;
+	margin-top: 2rem;
 }
-.header {
+.Hiringheader {
 	margin: auto 0;
 	position: relative;
 	justify-content: flex-start;
@@ -24,12 +26,10 @@
 	margin-bottom: 15rem;
 }
 #hire {
-	/* font-family: rosewood-std, sans-serif; */
 	font-weight: 400;
 	font-size: 4em;
 	font-style: normal;
 	color: #37d68c;
-	opacity: 87%;
 }
 
 .applyButton {
@@ -53,32 +53,44 @@
 	margin-right: 0.75em;
 }
 
-.cards {
-	/* background-image:linear-gradient(90deg,#d5ff45,#d5dd45); */
+/* .cards {
+	/* background-image:linear-gradient(90deg,#d5ff45,#d5dd45); 
 	width: 100%;
 	position: absolute;
-	top: 95%;
-	display: grid;
-	grid-template-columns: auto auto auto;
+	top: 65%;
+	display: flex;
+	/* grid-template-columns: auto auto auto; 
 	background-color: #37d68c;
 	overflow: hidden;
 	left: 0;
+	flex-wrap: wrap;
 	margin: 0 auto;
-	padding: 2rem;
+	/* padding: 2rem; 
+} */
+
+.cards{
+	background-color: #37d68c;
+	display: flex;
+	flex-wrap: wrap;
+	margin: 0 auto;
+	padding: 1rem;
 }
 
 .Cardsmain {
-	display: flex;
+	/* display: flex; */
 	--s: 14em; /* size  */
 	--m: 3.25em; /* margin */
 	--f: calc(1.732 * var(--s) + 4 * var(--m) - 1px);
 	background-color: #37d68c;
 	/* height: 100vh; */
-	gap: 30em;
+	/* gap: 30em; */
 	/* flex: 3; */
 	bottom: 0;
-	margin-left: 5rem;
+	/* margin-left: 5rem; */
 	/* overflow: hidden; */
+	flex-basis: 26.333333%;
+	gap: 2rem;
+	margin-left: 7rem;
 }
 
 .cardContainer {
@@ -86,9 +98,7 @@
 	cursor: pointer;
 	transform-style: preserve-3d;
 	transition: transform 0.5s;
-	/* grid-template-columns: 1fr 1fr 1fr; */
-	/* columns: 3; */
-	display: grid;
+	display: flex;
 }
 
 .cardContainer div {
@@ -118,12 +128,12 @@
 }
 
 .cardContainer:hover {
-	/* transform: scale(1.2); */
+	transform: scale(1.21);
 	/* transform: rotateY(180deg); */
 }
 
 .content:hover {
-	transform: translateY(-3px);
+	/* transform: translateY(-3px); */
 	/* transform: rotateY(-180deg); */
 }
 
@@ -148,92 +158,4 @@
 
 .hidden {
 	display: none;
-}
-
-.card {
-	position: relative;
-	width: 18.75em;
-	height: 21.688em;
-	font-family: helvetica, arial, sans-serif;
-	color: #d5ff45;
-	cursor: pointer;
-	transform-style: preserve-3d;
-	transition: transform 0.2s;
-	-moz-backface-visibility: hidden;
-}
-.card.flipped {
-	transform: rotateY(180deg);
-}
-
-.card.flipped.front {
-	z-index: 2;
-}
-
-.card.flipped.back {
-	z-index: 3;
-}
-
-.hexagon {
-	position: absolute;
-	width: 300px;
-	height: 173.21px;
-	background-color: #001b29;
-	margin: 86.6px 0;
-	box-shadow: 0 0 20px rgba(0, 0, 0, 0.6);
-	-webkit-backface-visibility: hidden;
-	-moz-backface-visibility: hidden;
-	-ms-backface-visibility: hidden;
-	transition: transform 0.5s;
-	z-index: 3;
-}
-
-.hexagon:hover {
-	transform: translateY(-3px);
-}
-
-.hexagon:before,
-.hexagon:after {
-	content: "";
-	position: absolute;
-	z-index: 1;
-	width: 212.13px;
-	height: 212.13px;
-	transform: scaleY(0.5774) rotate(-45deg);
-	background-color: inherit;
-	left: 43.934px;
-}
-
-.hexagon:before {
-	top: -106.066px;
-}
-
-.hexagon:after {
-	bottom: -106.066px;
-}
-
-.hexagon.back {
-	transform: rotateY(180deg);
-	background-color: #001b29;
-	color: rgb(239, 246, 209);
-	z-index: 2;
-}
-
-.card-content {
-	text-align: center;
-	position: relative;
-	z-index: 2;
-}
-.see-more {
-	/* position: relative; */
-	/* margin-bottom:calc(var(--s) * 0.25);  */
-	margin-left: calc(var(--s) * 0.3);
-	z-index: 2;
-	transition: color 0.2s, transform 0.2s;
-	font-size: 1.5em;
-	bottom: 3em;
-}
-
-.card:hover .see-more {
-	color: #fff;
-	transform: rotate(45deg);
 }
